+++ modules/account-lib/package.json
@@ -61,6 +61,7 @@
     "bs58": "^4.0.1",
     "bs58check": "^2.1.2",
     "casper-client-sdk": "1.0.39",
+    "create-hmac": "^1.1.7",
     "elliptic": "^6.5.2",
     "ethereumjs-abi": "^0.6.5",
     "ethereumjs-util": "6.2.1",
