+++ modules/account-lib/src/utils/ed25519KeyDeriver.ts
@@ -1,11 +1,7 @@
 /**
  * @prettier
- * @hidden
  */
-
-/**
- */
-import * as createHmac from 'create-hmac';
+import createHmac from 'create-hmac';
 
 export interface HdKeypair {
   key: Buffer;
@@ -47,7 +43,7 @@ export class Ed25519KeyDeriver {
     const segments = path.split('/').slice(1).map(this.replaceDerive);
     return segments.reduce(
       (parentKeys, segment) => this.CKDPriv(parentKeys, segment + Ed25519KeyDeriver.HARDENED_OFFSET),
-      { key, chainCode }
+      { key, chainCode },
     );
   }
